fix: correct the cron jobs
