+++ backend/core/src/shared/shared.module.ts
@@ -27,7 +27,7 @@ import Joi from "joi"
         TWILIO_PHONE_NUMBER: Joi.string().default("dummy_phone_number"),
         AUTH_LOCK_LOGIN_AFTER_FAILED_ATTEMPTS: Joi.number().default(5),
         AUTH_LOCK_LOGIN_COOLDOWN_MS: Joi.number().default(1000 * 60 * 30),
-        AFS_PROCESSING_CRON_STRING: Joi.string().default("0 0 * * *"),
+        AFS_PROCESSING_CRON_STRING: Joi.string().default("0 * * * *"), // Setting to every hour for testing purposes
       }),
     }),
   ],
